Move specs to core.h
